Add boto tests for InstanceInitiatedShutdownBehavior=terminate
The EC2 API supports two different behaviors when an instance
shuts itself down -- 'stop' and 'terminate'.  This boto test
starts an instance of each and uses ssh to call shutdown inside
the instance, and then checks to see if the instance ends
up in 'stopped' state or if it's gone completely.

Fixes bug #1131395

Change-Id: Ia79083379f8161b77336cf595c90f1dbffbf7409